Implement _.each which calls a function to each element of an array or object
Initially there should be a check to see if the collection is an array or an object. If it is an array you will use a for loop calling the iterator function on each element of the array with additional access to the current index or collection. If the array check is false then a for in loop will be used to call the iterator with each key in the collection while also providing access to each key and the object itself. Could be implemented using Object.keys(collection) and just a for loop; Remember to refactor this is there is time. In the _.idexOf function you can see why allowing access to the index or key of the collection within the iterator fnction is useful.