Implement _.filter which returns an array of elements that pass a truth test from an input array
First declare an array variable (filtered) which will be returned at the end. We can use each and an anonymous function to apply the test function to each element of the array. If that test is truthful then we .push the element into the filtered array. At the end we will return the array.