Implement _.reject which tests each element of an array with a truth test and returns an array of all false values
To reuse filter we need it to return an array of false values. This means we return filter and then replace the test function with a function that inverts the value of the test so that elements that don't pass the test return true and get pushed to the filtered array the will be returned.