Learn and implement _.uniq which removes duplicate values from an array
Uniq returns a duplicate free version of the input array. This is done by creating a new array and using the indexOf function to check the empty array if it contains each element of the input array and to .push the value if it doesn't. It does not change the input array.