Implement _.reduce and make the tests pass and understand implementation of contains
Reduce takes as arguments a collection, a function to be applied, and an optional start value. The iterator function takes in the start value and a collection element and returns a result which is the start value in the function which will be called on the next element. This reduces a collection into one result at the end. If a start value is not given it should use the first element as the start value. It should also call the function even if it will return undefined (not sure if I implemented that correctly).

Contains checks a collection to see if it contains a target value. If it does it returns true. It uses reduce and an anonymous operator function to see if at least one of the elements contains the target value. The start value is false and as soon as the equality check returns true the function keeps returning true until it ends.