Understand and implement _.every and make all the tests pass
Every passes each element of a collection through a test function which can return true or false. Using reduce and an initial value of true the elements run through and keep returning true until the test doesn't pass which will then return false or until the last element is reached (which means that every element in the collection passed the test). To make every work even when no callback is defined we set the default callback to _.identity just like the notes mention may be useful to do.