Understand and implement _.some and make all of the tests pass
Underscore some takes a collection and a test function and tests each element until one passes the truth test and returns true if none pass it returns false. We can use the inverse of the every function because that returns true until the truth test doesn't pass. We need to send the inverse result of the iterator function into the every function and invert that result. The only problem with this is that even if one element passes the test it will still iterate over the whole collection returning true. _.identity is still used as the fefault function.