Understand and implement _.extend & _.defaults and make all the tests pass
Underscore extends takes in multiple arguments of objects where the first object is the destination object and all following objects are the origin objects which contain properties. Each origin object keys and properties are assigned to the destination object. For defualts a check is done before assignment to see if that key exists in the destination object and only proceeds with assignment if the key is not present in the destination object.