Fix memoize as it was not working as intended, Implement _.delay
Memoize was still passing the tests by using Array.prototype.slice.apply(arguments,arguments) which allowed for the function to be called again but saved additional data in the memo. Overall using apply(arguments, arguments) is not good code. Found a better solution for storing argument lists in JSON.stringify.

Delay uses the setTimeout function and calls a function with the corresponding arguments after a certain amount of time.