Implement _.shuffle, undestrand it and make all tests pass
Shuffle does not modify the original array. Assuming every arguemnt that will be passed in we can use .slice on this array to copy it to a new array. Then we take the last element of the array and replace it with a randomly selected element from the rest of the array. As we move towards the first element in the array we keep randomly picking elements in front of it to replace the current element. This is known as a Fisher Yates shuffle.